Nilam/upload&download

+++ src/app/controllers/books.controller.js
@@ -1,6 +1,6 @@
 const db = require('../../utils/db.setup.util')
 const { book } = db.models
-const NotFoundError = require('../../errors/classes/sub/notFound.error');
+const { ForbiddenResourceError, NotFoundError } = require('../../errors/utils/errors.interface.util')
 
 module.exports = {
     get: async(req, res, next) => {
@@ -39,30 +39,9 @@ module.exports = {
     },
 
     insert: async(req, res, next) => {
+        if (res.locals.admin == false)
+            throw new ForbiddenResourceError('Forbidden Resource.')
 
-        if (!req.files || Object.keys(req.files).length === 0) {
-            return res.status(400).send('No files were uploaded.');
-        }
-
-        const file = req.files.file;
-        const fileSize = file.data.length;
-        const ext = path.extname(file.name);
-        const fileName = file.md5 + ext;
-        const url = `${req.protocol}://${req.get("host")}/images/${fileName}`;
-        const allowedType = ['.png', '.jpg', '.jpeg'];
-
-        if (!allowedType.includes(ext.toLowerCase())) return res.status(422).json({ msg: "Invalid Images" });
-        if (fileSize > 5000000) return res.status(422).json({ msg: "Image must be less than 5 MB" });
-
-        file.mv(`./public/images/${fileName}`, async(err) => {
-            if (err) return res.status(500).json({ msg: err.message });
-            try {
-                await Product.create({ name: name, image: fileName, url: url });
-                res.status(201).json({ msg: "Product Created Successfuly" });
-            } catch (error) {
-                console.log(error.message);
-            }
-        })
         try {
             const response = await book.create({
                 idBuku: req.body.idBuku,
@@ -89,6 +68,9 @@ module.exports = {
     },
 
     updateId: async(req, res, next) => {
+        if (res.locals.admin == false)
+            throw new ForbiddenResourceError('Forbidden Resource.')
+
         try {
             const idBuku = req.params.id
             const response = await book.update({
@@ -126,6 +108,9 @@ module.exports = {
     },
 
     remove: async(req, res, next) => {
+        if (res.locals.admin == false)
+            throw new ForbiddenResourceError('Forbidden Resource.')
+
         try {
             const response = await book.destroy({
                 where: {
